my code review
