+++ app/src/main/java/com/uc/mymovies/service/UserService.kt
@@ -1,4 +1,5 @@
 package com.uc.mymovies.service
 
 class UserService {
+
 }
